Solution

+++ src/main/java/com/kris/javaYlab/taskPell/Pell.java
@@ -9,8 +9,8 @@ public class Pell {
         try (Scanner scanner = new Scanner(System.in)) {
             int n = scanner.nextInt();
             if (n >= 0 && n <= 30) {
-                double pell = ((Math.pow(1 + (sqrt(2)), n)) - (Math.pow(1 - (sqrt(2)), n))) / ((sqrt(2)) + (sqrt(2))) + 1;
-                System.out.println((int) pell);
+                int pell = (int) (((Math.pow((1 + (sqrt(2))), n)) - (Math.pow((1 - (sqrt(2))), n))) / ((sqrt(2)) + (sqrt(2))) + 1);
+                System.out.println(pell);
             }
         }
     }
